aanpassen price type naar Decimal toevoegen prijzen aan producten in ProductRepository

+++ Grocery.Core/Models/Product.cs
@@ -8,12 +8,12 @@ namespace Grocery.Core.Models
         public int stock;
 
         [ObservableProperty]
-        public int price;
+        public Decimal price;
         public DateOnly ShelfLife { get; set; }
         public Product(int id, string name, int stock) : this(id, name, stock, default, default) { }
         public Product(int id, string name, int stock, DateOnly shelfLife) : this(id, name, stock, shelfLife, default) { }
 
-        public Product(int id, string name, int stock, DateOnly shelfLife, int price) : base(id, name) 
+        public Product(int id, string name, int stock, DateOnly shelfLife, Decimal price) : base(id, name) 
         {
             Stock = stock;
             ShelfLife = shelfLife;
